pyproject.toml
[tool.poetry]
name = "hogger"
version = "0.1.0"
description = ""
authors = ["Nick Saccente"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.3.0"
mysql-connector-python = "^8.1.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
